Add files via upload
The retention and acquisition of users are the major concern in telecom industry .Churn prediction contributes to identify those users who are likely to switch a company over other. Through this project I try to make a good machine learning model to predict the customer churn and also try to discover the some underlying relationship between different features of the customers that leads to churning.